Refactored the structure of the tests.
