+++ tests/config/test_array_expansion.py
