+++ tests/conftest.py
@@ -1,6 +1,12 @@
 import builtins
 import sys
+from pathlib import Path
 import pytest
+
+PROJECT_ROOT = Path(__file__).resolve().parent.parent
+SRC_ROOT = PROJECT_ROOT / "src"
+if str(SRC_ROOT) not in sys.path:
+    sys.path.insert(0, str(SRC_ROOT))
 
 
 @pytest.fixture(scope="session", autouse=True)
