--- tests/test_my_module.py
@@ -1,30 +1,0 @@
-"""Tests for the tud_lbm package."""
-
-import pytest
-
-
-@pytest.mark.skip(reason="Package initialization needs to be completed")
-def test_package_imports():
-    """Test that the tud_lbm package can be imported.
-
-    Given: tud_lbm package is installed
-    When: attempting to import tud_lbm
-    Then: import should succeed
-    """
-    import tud_lbm  # noqa: F401
-
-    assert True, "Package imported successfully"
-
-
-@pytest.mark.skip(reason="Package initialization needs to be completed")
-def test_package_has_core_modules():
-    """Test that runner modules exist in the package.
-
-    Given: tud_lbm package is imported
-    When: checking for runner submodules
-    Then: should have expected submodules
-    """
-    import tud_lbm
-
-    # Verify package structure exists
-    assert hasattr(tud_lbm, "__file__"), "Package should have __file__ attribute"
