+++ tests/util/plotting/test_plotting.py
@@ -1,13 +1,8 @@
 """Tests for the plotting operator package and figure builder."""
 
 from __future__ import annotations
-import sys
-from pathlib import Path
 import numpy as np
 import pytest
-
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
-
 from config import SimulationConfig
 from registry import get_operator_names
 from util.plotting import visualise
